Add script to bring up containers in order
Update docker_files. Containers only depend on mongo
