+++ experiments/docker_files/exp-1-approach-4.yaml
@@ -3,10 +3,10 @@ services:
     command:
     - python3
     - ccu.py
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - non_intentional_delays
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: ccu
     depends_on:
     - mongo
@@ -23,29 +23,16 @@ services:
     - 27017:27017
     volumes:
     - /data/db:/data/db
-  mrta_experiment:
+  mrta:
     build:
       context: ../../
       dockerfile: Dockerfile
     command:
     - python3
-    - run.py
+    - experiment.py
     - non_intentional_delays
     - tessi-corrective-re-allocate
-    container_name: mrta-experiment
-    depends_on:
-    - mongo
-    - ccu
-    - robot_001
-    - robot_proxy_001
-    - robot_002
-    - robot_proxy_002
-    - robot_003
-    - robot_proxy_003
-    - robot_004
-    - robot_proxy_004
-    - robot_005
-    - robot_proxy_005
+    container_name: mrta
     image: ropod-mrs
     network_mode: host
     stdin_open: 'true'
@@ -56,10 +43,10 @@ services:
     - python3
     - robot.py
     - robot_001
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - non_intentional_delays
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: robot_001
     depends_on:
     - mongo
@@ -73,10 +60,10 @@ services:
     - python3
     - robot.py
     - robot_002
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - non_intentional_delays
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: robot_002
     depends_on:
     - mongo
@@ -90,10 +77,10 @@ services:
     - python3
     - robot.py
     - robot_003
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - non_intentional_delays
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: robot_003
     depends_on:
     - mongo
@@ -107,10 +94,10 @@ services:
     - python3
     - robot.py
     - robot_004
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - non_intentional_delays
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: robot_004
     depends_on:
     - mongo
@@ -124,10 +111,10 @@ services:
     - python3
     - robot.py
     - robot_005
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - non_intentional_delays
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: robot_005
     depends_on:
     - mongo
@@ -141,10 +128,10 @@ services:
     - python3
     - robot_proxy.py
     - robot_001
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - non_intentional_delays
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: robot_proxy_001
     depends_on:
     - mongo
@@ -158,10 +145,10 @@ services:
     - python3
     - robot_proxy.py
     - robot_002
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - non_intentional_delays
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: robot_proxy_002
     depends_on:
     - mongo
@@ -175,10 +162,10 @@ services:
     - python3
     - robot_proxy.py
     - robot_003
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - non_intentional_delays
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: robot_proxy_003
     depends_on:
     - mongo
@@ -192,10 +179,10 @@ services:
     - python3
     - robot_proxy.py
     - robot_004
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - non_intentional_delays
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: robot_proxy_004
     depends_on:
     - mongo
@@ -209,10 +196,10 @@ services:
     - python3
     - robot_proxy.py
     - robot_005
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - non_intentional_delays
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: robot_proxy_005
     depends_on:
     - mongo
