+++ experiments/docker_files/exp-4-1-approach-4.yaml
@@ -3,10 +3,10 @@ services:
     command:
     - python3
     - ccu.py
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - robot_scalability_1
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: ccu
     depends_on:
     - mongo
@@ -23,21 +23,16 @@ services:
     - 27017:27017
     volumes:
     - /data/db:/data/db
-  mrta_experiment:
+  mrta:
     build:
       context: ../../
       dockerfile: Dockerfile
     command:
     - python3
-    - run.py
+    - experiment.py
     - robot_scalability_1
     - tessi-corrective-re-allocate
-    container_name: mrta-experiment
-    depends_on:
-    - mongo
-    - ccu
-    - robot_001
-    - robot_proxy_001
+    container_name: mrta
     image: ropod-mrs
     network_mode: host
     stdin_open: 'true'
@@ -48,10 +43,10 @@ services:
     - python3
     - robot.py
     - robot_001
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - robot_scalability_1
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: robot_001
     depends_on:
     - mongo
@@ -65,10 +60,10 @@ services:
     - python3
     - robot_proxy.py
     - robot_001
-    - --approach
-    - tessi-corrective-re-allocate
     - --experiment
     - robot_scalability_1
+    - --approach
+    - tessi-corrective-re-allocate
     container_name: robot_proxy_001
     depends_on:
     - mongo
