+++ experiments/docker_files/exp-4-4-approach-14.yaml
@@ -3,10 +3,10 @@ services:
     command:
     - python3
     - ccu.py
-    - --approach
-    - tessi-dsc-corrective-re-allocate
     - --experiment
     - robot_scalability_4
+    - --approach
+    - tessi-dsc-corrective-re-allocate
     container_name: ccu
     depends_on:
     - mongo
@@ -23,27 +23,16 @@ services:
     - 27017:27017
     volumes:
     - /data/db:/data/db
-  mrta_experiment:
+  mrta:
     build:
       context: ../../
       dockerfile: Dockerfile
     command:
     - python3
-    - run.py
+    - experiment.py
     - robot_scalability_4
     - tessi-dsc-corrective-re-allocate
-    container_name: mrta-experiment
-    depends_on:
-    - mongo
-    - ccu
-    - robot_001
-    - robot_proxy_001
-    - robot_002
-    - robot_proxy_002
-    - robot_003
-    - robot_proxy_003
-    - robot_004
-    - robot_proxy_004
+    container_name: mrta
     image: ropod-mrs
     network_mode: host
     stdin_open: 'true'
@@ -54,10 +43,10 @@ services:
     - python3
     - robot.py
     - robot_001
-    - --approach
-    - tessi-dsc-corrective-re-allocate
     - --experiment
     - robot_scalability_4
+    - --approach
+    - tessi-dsc-corrective-re-allocate
     container_name: robot_001
     depends_on:
     - mongo
@@ -71,10 +60,10 @@ services:
     - python3
     - robot.py
     - robot_002
-    - --approach
-    - tessi-dsc-corrective-re-allocate
     - --experiment
     - robot_scalability_4
+    - --approach
+    - tessi-dsc-corrective-re-allocate
     container_name: robot_002
     depends_on:
     - mongo
@@ -88,10 +77,10 @@ services:
     - python3
     - robot.py
     - robot_003
-    - --approach
-    - tessi-dsc-corrective-re-allocate
     - --experiment
     - robot_scalability_4
+    - --approach
+    - tessi-dsc-corrective-re-allocate
     container_name: robot_003
     depends_on:
     - mongo
@@ -105,10 +94,10 @@ services:
     - python3
     - robot.py
     - robot_004
-    - --approach
-    - tessi-dsc-corrective-re-allocate
     - --experiment
     - robot_scalability_4
+    - --approach
+    - tessi-dsc-corrective-re-allocate
     container_name: robot_004
     depends_on:
     - mongo
@@ -122,10 +111,10 @@ services:
     - python3
     - robot_proxy.py
     - robot_001
-    - --approach
-    - tessi-dsc-corrective-re-allocate
     - --experiment
     - robot_scalability_4
+    - --approach
+    - tessi-dsc-corrective-re-allocate
     container_name: robot_proxy_001
     depends_on:
     - mongo
@@ -139,10 +128,10 @@ services:
     - python3
     - robot_proxy.py
     - robot_002
-    - --approach
-    - tessi-dsc-corrective-re-allocate
     - --experiment
     - robot_scalability_4
+    - --approach
+    - tessi-dsc-corrective-re-allocate
     container_name: robot_proxy_002
     depends_on:
     - mongo
@@ -156,10 +145,10 @@ services:
     - python3
     - robot_proxy.py
     - robot_003
-    - --approach
-    - tessi-dsc-corrective-re-allocate
     - --experiment
     - robot_scalability_4
+    - --approach
+    - tessi-dsc-corrective-re-allocate
     container_name: robot_proxy_003
     depends_on:
     - mongo
@@ -173,10 +162,10 @@ services:
     - python3
     - robot_proxy.py
     - robot_004
-    - --approach
-    - tessi-dsc-corrective-re-allocate
     - --experiment
     - robot_scalability_4
+    - --approach
+    - tessi-dsc-corrective-re-allocate
     container_name: robot_proxy_004
     depends_on:
     - mongo
